% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_exon_clusters.R
\name{make_exon_clusters}
\alias{make_exon_clusters}
\title{Create Exon Clusters for a Species}
\usage{
make_exon_clusters(species = "hg38", min_gap_width = 10, verbose = TRUE)
}
\arguments{
\item{species}{Species genome assembly. Options: "hg38" (human, default),
"mm10" (mouse), "rn6" (rat), "dm6" (fly), "ce11" (worm).
Must have corresponding TxDb package installed (e.g., TxDb.Hsapiens.UCSC.hg38.knownGene).}

\item{min_gap_width}{Integer. Exons separated by gap <= this distance are merged
into the same cluster. Default: 10. Set to \code{NA} to only merge overlapping exons.}

\item{verbose}{Print progress messages? Default: TRUE.}
}
\value{
A \code{GRanges} object with exon clusters. The metadata column
\code{revmap} contains indices of original exons that were merged into each cluster.
}
\description{
Fetches exon coordinates from TxDb for the specified species and
creates exon clusters by merging overlapping/adjacent exons using
\code{\link[GenomicRanges:reduce]{GenomicRanges::reduce()}}.
}
\examples{
\dontrun{
# Human hg38 with default min_gap_width=10
clusters <- make_exon_clusters(species = "hg38")

# Mouse mm10 with stricter clustering
clusters <- make_exon_clusters(species = "mm10", min_gap_width = 5)

# Only merge overlapping exons
clusters <- make_exon_clusters("hg38", min_gap_width = NA)
}
}
