% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_exon_reads_hts.R
\name{extract_exon_reads_hts}
\alias{extract_exon_reads_hts}
\title{htslib-core extractor (fast path)}
\usage{
extract_exon_reads_hts(
  bam,
  chr,
  start,
  end,
  out_bam,
  tsv,
  xf_values = c(25L, 17L)
)
}
\arguments{
\item{bam}{Path to input BAM/CRAM}

\item{chr}{Region chrom/contig (e.g., "chr1")}

\item{start}{1-based inclusive}

\item{end}{1-based inclusive}

\item{out_bam}{Path for filtered BAM to write (or NULL to skip)}

\item{tsv}{Path to TSV to write}

\item{xf_values}{integer vector; default c(25L,17L)}
}
\description{
htslib-core extractor (fast path)
}
